requirements.txt
adjacent==1.0.0
beautifulsoup4==4.6.0
cent==2.1.0
certifi==2017.7.27.1
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.4.2
Django==1.11.3
django-allauth==0.32.0
django-appconf==1.0.2
django-autocomplete-light==3.2.9
django-contrib-comments==1.8.0
django-crispy-forms==1.6.1
django-el-pagination==3.1.0
django-endless-pagination==2.0
django-feedback-form==1.3
django-filter==1.0.4
django-filters==0.2.1
django-fluent-comments==1.4.2
django-friendship==1.5.0
django-inplaceedit==1.4.1
django-libs==1.67.12
django-likes==1.11
django-mailer==1.2.2
django-model-utils==3.0.0
django-modelcluster==3.1
django-notifications-hq==1.3
django-redis==4.8.0
django-redis-cache==1.7.1
django-redis-sessions==0.5.6
django-rest-auth==0.9.1
django-secretballot==0.7.0
django-tag-parser==3.1
django-taggit==0.22.1
django-treebeard==4.1.2
django-webpack-loader==0.5.0
djangorestframework==3.6.3
djcommon==1.1.1
GDAL==2.1.4
gunicorn==19.7.1
html5lib==0.999999999
idna==2.5
jsonfield==2.0.2
lockfile==0.12.2
Markdown==2.6.8
oauthlib==2.0.2
olefile==0.44
optional-django==0.3.0
Pillow==4.2.1
psycopg2==2.7.3
python-akismet==0.3.0
python3-openid==3.1.0
pytz==2017.2
redis==2.10.5
requests==2.18.2
requests-oauthlib==0.8.0
simplejson==3.11.1
six==1.10.0
Unidecode==0.4.21
urllib3==1.22
wagtail==1.11.1
webencodings==0.5.1
Willow==0.4